refactoring + revert .env.dev
