+++ .env.dev
@@ -1,2 +1,2 @@
-API_URL= "http://192.168.43.160:3000/"
+API_URL="http://smtp-dev-env.eba-5jqrxjhz.eu-west-3.elasticbeanstalk.com/"
 VERSION="development"
