+++ android/app/src/main/java/com/smtp/smtp/navigation/Pause.java
@@ -1,4 +1,4 @@
-package com.smtp.smtp;
+package com.smtp.smtp.navigation;
 
 import android.content.Context;
 import android.util.Log;
@@ -11,6 +11,8 @@ import com.android.volley.VolleyError;
 import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.HttpHeaderParser;
 import com.android.volley.toolbox.StringRequest;
+import com.smtp.smtp.BuildConfig;
+import com.smtp.smtp.http.RequestManager;
 
 import org.json.JSONObject;
